store: phobos_locate
We add the phobos_locate call to the phobos_store.h API.

This call retrieves the hostname from which a given object is
available. We could then execute a phobos_get on this hostname.

This call is implemented by calling a new layout_locate call which
looks at the media hosting extents of an object.

Change-Id: I1b8a7c0527a6570e39630231f8297462fdd206ce
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/8485
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>