rfc: remove id field from pho_lock data structure
This field is only used in the DSS context and does not need for now to
be known outside this context.

The daemon used this field to optimize the distinction between internal
and external locks. Instead, a boolean field is added to the structure
to maintain this optimization.

Also, pho_lock util functions are added for convenience.

Change-Id: I880172a12cf78273d4e9c415542b375e6c2081fc
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/8674
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>