lrs: lock when loading device
Lock are now "permanent" on device and media. When a lrs starts,
it takes a lock on all devices that are registered to this host
into the DSS. These locks on devices are released when the LRS
stops. As soon as a medium is loaded into a device locked by a LRS,
the LRS locks this medium and releases it when the medium is
unloaded.

With this new "permanent" lock policy, the locate feature can
efficiently use existing locks to find the host currently owning a
resource.

Change-Id: I47d207dcdb3f31479a3380e284cf97ea01a2ae4a
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/8820
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>