tests: refactor acceptance.sh
The goal of this stack is to offer a testsuite which can be used both in
the non-regression process and on pre-production systems.

At the end of this stack, the old acceptance.sh will be removed, and
parts which will not be maintained in the pre-production testsuite will
be added in new tests.

Change-Id: I60803ad9d24a7f092f71f0e6452f1632a82a729c
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/9091
Reviewed-by: Linter
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>