dss: dss_device_update_adm_status
By updating only the adm_status column of the device table, we
have an atomic update and we don't need any locking.

We split the dss_device_set function in dss_device_insert,
dss_device_delete and dss_device_update_adm_status to emphasize the
specific behaviour of the dss_device_update_adm_status call which
is not a generic update call anymore.

Change-Id: I9a0222b496d207ec95ac040ff02c6d6f92448f18
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/9132
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>