tests: split raid1_split, move raid1_simple_split
This patch splits raid1_split.sh into two, with the first
part going into test_raid1_split.sh and the second staying
in the same file but renamed to test_raid1_split_locate.sh.

This is done because raid1_split.sh first checks that a
put/get with a replica count of 2 works properly, which
is already the goal of another test (though it tests only
with replica count of 1). So the two are merged.

The second part of raid1_split.sh tests the locate feature,
so the file is renamed to be more explicit.

Change-Id: Iff72a1a4ac42b579c29727c4b53b6612bee149cb
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/9169
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>