bugfix: fix test_media incomplete setup
This patch fixes an issue that provoke test_media's failure
if test_store (through the test test_layouts) is not used
before. When test_store is launched, it will create 5
directories that will be used as media. Those directories
are not removed when the test ends, and test_media was using
those same directories for itself.

Therefore, this patch fixes test_media's incomplete setup and
test_store's incomplete cleanup.

Change-Id: I2ad04686f8839725785df9a889d300e62f2ff5a4
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/9190
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>