io: IO adapter provides preferred I/O size
This adds a new call to I/O adapters: ioa_preferred_io_size()
which indicates the preferred I/O size for the filesystem.
This call is called by the layout before performing the I/O.

On POSIX FS, the returned value is the one from statfs.f_bsize
(e.g. 4K on a ext4 file system).
On LTFS, this value contains the optimal value for the given model of tape
(e.g. 512K on LTO6).

This can be overridden by specifying a value in the configuration
(parameter: io::io_block_size). In this case, the specified value
is used for all filesystems managed by Phobos.

Change-Id: I02944f8d0b7d4c41c7241bd1dcfb4658019a4bc0
Signed-off-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/9871
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>