lrs: add the structure lrs_dev_hdl
The goal of this patch is to start moving elements related to the
device from lrs_sched.c to lrs_device.c.

The structure lrs_dev_hdl will provide a simple API to handle devices
at the lrs_sched level. This patch only introduces the structure and
the basic functions it needs.

The lrs_dev structure has also been added in order to properly
isolate distinct elements of the device. It will replace dev_descr
in the following patches.

The following patches will start using this strucure and move the
responsability of manipulating device related operations from
lrs_sched to lrs_dev.

Change-Id: Ib3cb612e54ad32ffda13c8b6c79f229ca22c3057
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/9796
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>