cli: change drive lock API
One of the new behaviors of phobos daemon is to lock every drive it
possesses, which makes the `phobos drive lock` command useless without
the `--force` option.

The new API will consider the old `--force` option as default, meaning
that drives will be locked even if a daemon already uses it, and will
instead offer to wait for the daemon acknowledgement or not.

Change-Id: I0affed6caa65aeec5c7ce9bfc152c063de43b82b
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10362
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>