tests: fix timeout issue
The 'timeout' print is due to the script which prevents multiple
instances of `make check` to be launched concurrently. It creates a
background process which waits until the global test lock file is
removed or a deadline is reached to remove this lock file.

This script was made 7 years ago and considered our tests will not last
more than 5 minutes. This value is increased to 30 minutes, due to the
expansion of our tests.

Also, the script was called two times, once for python internals and
another one for c internals and extern tests. Now, the script is only
called once, when `make check` starts.

Change-Id: I9d88163c76bf985877dcb425b118fd687b0756f6
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10241
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>