lrs,cli: improve error management during drive status
On the client side, we make sure that every element in the JSON
object is either there or has a default empty value.

On the LRS side, we don't propagate errors during the building
of the JSON object since a partial information may still be of
use to the caller.

Change-Id: I566c563ec7b40b0ee5b0df06e17a0c1c3fb69457
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10244
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>