lrs: dev_handle_read_write in device thread
If the device faces an error when preparing its sub_request and
there is a chance to execute it again (ie: a write or a read with
an error while there are other candidate media in this read
allocation), the device thread requeue the sub request into the
scheduler.

If a device should run a sub_request of a definitly failed request,
it cancels it.

From device thread's point of view, preparing a read or preparing a
write needs exactly same actions. The device thread only need to
load and mount the right medium.

Change-Id: I5a58a066a77e8fb5227976cb15993aad25f708fc
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10425
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>