ltfs: stop killing ltfs process on unmount
Tests on Centos8 failed because tapes could not be unloaded. Further
investigations show that the script we call when unmounting a tape kill
the ltfs process prior/without doing the synchronization, which lead to
not remove the 'prevent medium removal' lock.

As an unmount will still stop the ltfs process, we do not need to kill
it.

Change-Id: Iffbe5c906a8127bc94c8d20289f8487e17681a09
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10382
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>