lib adapters: modularize the lib adapter
Change the way the lib adapters are managed to dlopen
them when required, instead of having a specific
reference to one structure, mimicking how the layouts
are managed.

Currently only opens the lib_adapter for dummy devices.
That will be changed next patch to do the same for the
scsi adapter.

Also create the "ldm-modules" directory that will contain
said lib adapters.

Change-Id: Idd95bdc0ec1a71d56342bce89481a46b856bd8b3
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10516
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>