tests: test sync to stopped a device and replace controlled_store.c
In addition of testing ld_ongoing_io, we also test if a device
thread has waiting sync before declarating it "stopped".

"controlled_store.c" was launched in background to send a rwalloc
and receive the answer before waiting a signal to send the release
and receiving the release answer. But, the tests had no garantee
than after executing controlled_store in background the rwalloc
was ending before continuing test.

We replace it by a binary executed in two steps. The end of
execution of the first step garantees that the rwalloc is ended
and that we know wait for the release.

Change-Id: Ife6c8d772d63e8302ef55417c54ad3e2ce2c1edf
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10543
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>