lrs: device thread manages ongoing_io flag
Device ongoing_io flag is now set by the device thread itself
instead of the scheduler thread. This change clarifies the
concurrency between all threads.
This flag is now:
- read by the scheduler to know the status of the devices,
- set by the device when it deals with a new subrequest,
- cleared by the main lrs communication thread when a release is
received.

This modification also allows to differentiate the fact that a
device has a scheduled subrequest from the fact that the device
has an ongoing io.

A new "ld_ongoing_scheduled" flag is added to a device. This new
flag is managed by the scheduler thread. The scheduler thread
set it to true when it books a device for an IO as long as the
subrequest is not yet pushed to the device. If the scheduler
thread faces an error in next device allocations of the same
request, the device is "released" and the scheduler thread sets
back the ld_ongoing_scheduled flag to false. If the scheduler
successfully allocates the whole request, it sets back to false
the ld_ongoing_scheduled flag of each device when it pushes the
subrequest to the device thread.

Change-Id: I9e3dc2c33028927433f3c6a95c6e4e3c3da74faa
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10644
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>