test: make test_daemon more robust
If a daemon was still running when executing test_invalid_lock_file,
pgrep would catch this daemon and the test would fail.

Since the lock file is setup after the fork, the PID of the child is
stored in the PID file regardless of whether the lock file is created or
not. So the variable PID_DAEMON initialized by invoke_daemon can be used
to check for the specific child process.

Also add execute permission to test_lock_clean.sh

Change-Id: I9870745edf04f508bf184e1ab38fe0c1770b6cf9
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10770
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>