design: early locking at locate
We want to introduce early locking at locate to efficiently
redirect massive bunch of get requests sharing same media to the
same node.

This early lock will be executed by the client during the locate
call with a hostname potentially different from the current one.

We will need a new dss_lock_hostname call to be able to take a
lock with a hostname different from the current one.

Change-Id: I6aaa1eaba4372803772b5f3134a165706aca4327
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10842
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>