lrs: fix sync enqueue race condition between comm and device
When a sync request is successfully enqueued to device thread,
the comm thread don't need to lock the request to manage a
nonexistent error.

This patch fixes a race condition between the comm and the device
threads. Indeed, as soon as the sync request is queued, the device
thread could successfully run and free this request. That's
why the comm thread must not try to lock this request after having
successfully pushed it.

Change-Id: Ife637706bef6730df056ba1d2a7424c88c6fadad
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10861
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>