lrs: fix double free on lookup error
A lookup error can return EBUSY. The function dev_load
can therefore return EBUSY on lookup and move. If the error
happened on lookup, the medium_info we were trying to lookup
was freed but the caller interpreted EBUSY as "I can try again
later". This meant that an EBUSY error on lookup would result in
a use after free and possibly a double free if EBUSY was returned
two times in a row by a lookup on the same medium.

Change-Id: Ibf639bed1d1b12f68f08a87a691a36d6c1cc22da
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11064
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>