dss: replace PQescapeString() deprecated call
The caller will now use PQescapeStringConn(), which is threadsafe, as
described below:

'PQescapeString can be used safely in client programs that work with
only one PostgreSQL connection at a time (in this case it can find out
what it needs to know “behind the scenes”). In other contexts it is a
security hazard and should be avoided in favor of PQescapeStringConn.'

To this extent, the dss_handle is added to the saj structure.

Change-Id: Iea1f1bf95675af3f0e13e8d8cdeaf79a5ce397f3
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11087
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>