modules: fix bad management of global variables in modules
When compiling module libraries, global symbols such as phobos_log_level
where resolved inside the library. This means that pho_info would look
at the value of phobos_log_level local to the library and not to the one
defined in lrs_init for example. This means that when loading a module,
it would get the default value of the global variable which was
PHO_LOG_DEFAULT (e.g. info). Therefore, there was an inconsistency in
log levels between the main executable and modules. This was also true
for the content of the configuration file. The global pointer
config->cfg_items was NULL by default which means that config_is_loaded
would always return false and pho_cfg_get_val_from_level would always
return -ENODATA for PHO_CFG_LEVEL_PROCESS. Therefore, modules would not
have access to the parameters defined in the configuration file.

This patch introduces a new global structure phobos_global_context. It
is allocated and initialized with default values in phobos_init() and
released in phobos_fini(). phobos_init() will have to be called before
any phobos specific function and phobos_fini() when we are done working
with Phobos. A pointer to this global context will have to be passed to
each module when calling load_module().

Change-Id: I5f0c37bca50f218213d693720469a727a38c28a8
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11121
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>