io_sched: properly handle the device model
The technology of a drive can be infered from its model.
We first have to list all the technologies that can be handled
by Phobos ('supported_list' in the section 'tape_model').
Then, we look for the model of the drive ('models' in
'drive_type "<techno>_drive"').

As this process is complex, the result is cached in
lrs_dev::ld_technology.

Change-Id: Ib97c9daac481e824b9e0f30a412cb466a3ab7c73
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11172
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>