bugfix: fix umount error after systemctl stop on phobosd
When a "systemctl stop" is used, it sends (by default) a
SIGTERM to all processes in the control group, which
includes all forked processes (see [1]).

As such, by using this command on phobosd, it would send
a SIGTERM to the main process and all forked LTFS
processes, which mean the umounts started at the end of
the LRS would be terminated prematurely, which would
result in a SCSI error.

This patch changes the default behaviour of sending a
SIGTERM to all the control group by specifying that on a
"systemctl stop", the SIGTERM should only be sent to the
main process, then a SIGKILL will be sent to the whole
control group after 5 minutes (see [2]).

[1]: https://man7.org/linux/man-pages/man7/cgroups.7.html).
[2]: https://www.freedesktop.org/software/systemd/man/systemd.kill.html

Change-Id: Ia922766c8df32a43fd19423b5f944675700c3ce3
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11227
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>