admin,lrs: add "phobos sched fair_share" command
This command will enable an admin to dynamically update
the fair_share configuration of the daemon so that these
parameters can be changed depending on the current load of
the system.

This patch adds a function phobos_admin_sched_conf_set to the admin API.
This function could be used in the future to update any parameter of the
daemon.

Change-Id: I30fb8b11e00174ff47b09e254a2dc03801320833
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11265
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>