admin,lrs: display current fair_share config
If no min and no max argument are given to "phobos sched fair_share",
we send a request to fetch the configuration from the LRS and display
them.

This patch adds a function phobos_admin_sched_conf_get to the admin API.
This function could be used in the future to get any parameter of the
daemon.

Change-Id: I1455c93f2156e458305100097c8855690e496cbf
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11266
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>