io-sched: fix fair_share when a device is already allocated
When a device is already allocated to a scheduler, if we add one new
device, the previous implementation would only have one new device to
give. But since the scheduler with the biggest weight should have two,
this would result in an invalid memory access of the second index of the
device list of size 1.

Change-Id: Ia4d059b4a5c4ca2b89ab46c93d82385a6d74e7e4
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11384
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>