locate: add tape/drive compat check
This patch introduce a compatibility check when locating an object. When
doing so, if the medium is usable, not locked, and is a tape, we will
check its model against every drives a host has available. If the check
is successful, we consider the host has access to that medium.

Change-Id: I691ab0b68ac0e4f85db975a9b1a9ed0cb074f976
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11308
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>