compatibility: move tape_drive_compat function
This patch moves the "tape_drive_compat" function in the LRS scheduler
to a new file in the CFGmodule, as it relies on the configuration file
to check the compatibility.

The goal is to prepare for the following patch, which checks during
locate if a host has drives compatible to read the extents of an object.

Change-Id: I3d88f8463296669fb63ec87885020489ac6ee923
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11307
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>