locate: retrieve every host before locating
This patch modifes the locate feature to retrieve every host in the DSS
with an unlocked device and their list of devices before trying to
locate an object. This ensures that every host is considered when
locating, even those that do not possess a lock on an extent or are not
focus-host.

Change-Id: Ifc0f9263b4e580367fa6ace34d8939b0c956e5a1
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11359
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>