bugfix: fix invalid counting of nb_locked_splits
This patch fixes an issue with the counting of nb_locked_splits when
locating an object. For instance, before this patch, the following would
occur: consider an object divided in 2 splits with a replica count of 4.
For the first split, "vm1" has locked 3 of the medium, while "vm0" has
locked one. For the second split, "vm0" also has one lock, and the other
3 medium are unlocked.

| split 1 | split 2 |
---------------------
| [ vm0 ] | [ vm0 ] |
| [ vm1 ] | [     ] |
| [ vm1 ] | [     ] |
| [ vm1 ] | [     ] |

In this scenario, "vm1" has the biggest number of locks overall (3 on
split 1 and 0 on split 2) whereas "vm0" has only 2 locks (1 for split 1
and 1 for split2). Before this patch, since "vm1" had 3 locks on a
replica of the split, its "nb_locked_splits" value would be 3, while
"vm0"'s one would be 2. As such, "vm1" would be chosen because both
hosts can access the object and it has a higher "nb_locked_splits"
count. However, in this case, "vm0" can access 2 splits without needing
any other lock, while "vm1" would require an extra one on any of the
free media.

This patch changes that behaviour so that nb_locked_splits for any host
can increase by at most one in any single split.

Also add tests for verifying the corresponding issue.

Change-Id: I77ee1fc54df2bcc9caf64304269089d3e9e48bf8
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11360
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>