io_sched: add the possibility for a scheduler to ask for a device
In some situations, a scheduler needs a device that is
owned by another scheduler (e.g. we need to read an object
stored on a tape owned by the write scheduler).

In this case, the I/O scheduler implementation can call
io_sched_claim_device using the option IO_SCHED_EXCHANGE
to exchange one of its free devices with the desired device.

Change-Id: I3ce8e9e78cbbab52390c1772fc2c6ffb7f1552eb
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11385
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>