lrs: take client I/O error into account
When a client sends a release request with an error, we will
stop using the medium. A call to ldm_fs_df will be executed.
If the client encounters an error due to a lack of space on
the medium, the status will be updated accordingly.

Since the LRS is supposed to give a medium with enough space
for a write, this should not happen in practice. Therefore,
the medium and device will be set to failed.

Change-Id: I2963c73a0317e01581c4876e0febc94eef1cf141
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11402
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>