io-sched: add round robin algorithm to choose an I/O scheduler
When asking for the next request, the io_sched API will use the
io_sched_handle::next_request algorithm. The current algorithm was FIFO
which means that the next request is the oldest of each scheduler's next
request. This behavior can lead to a dead lock when using the fair_share
device dispatch algorithm.

Say a tape where an object we want to read is written belongs to the
format I/O scheduler. But the corresponding read request is older than
the format request. With the current fifo_next_request algorithm, we
will never schedule the format request since the read one will always be
chosen. But it can't be scheduled since the tape belongs to the format
scheduler. In this case, the format request should be scheduled first.

A round robin approach is more appropriate in the case of the fair share
algorithm. We want to try to schedule requests from each scheduler
independantly. Furthermore, the fair share algorithm aims at making each
type of request executed fairly. A fifo algorithm imposes a priority
between schedulers based on their capacity to handle requests quickly.
This means that a slow scheduler will prevent quicker ones from
scheduling their requests. From this perspective, it also makes sense
to not use fifo_next_request.

Change-Id: I26c1dbf3268234b917b4b4af25a3f58d697e817b
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11445
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>