io-sched: add a retry callback to the I/O scheduling API
The previous I/O scheduling API would call get_device_medium_pair
with a boolean `is_error` set to false when first trying to allocate the
request and then to true when a sub request was sent from a device
thread to the scheduler thread through the `retry_queue`.

The issue with this behavior was that a request could be sent back to
the scheduler if the error encountered was not related to the medium
(eg. `sub_request::failure_on_medium` was false). In that case, we can
reuse the medium targeted by the sub request or try a new one. But since
this information was stored in the sub request, the I/O scheduler had no
idea that this was the case. It assumed that, since this is a retry,
other media can be used in the request which lead to an invalid read in
the med_ids table of the read request.

To solve this issue, the error handling is now managed by a new callback
`io_scheduler_ops::retry`. This callback will take the whole sub request
as an argument. This means that the scheduler will now have all the
information the device thread had when it decided to retry the sub
request.

find_device_medium_on_error is renamed grouped_retry in the grouped_read
I/O scheduler.

Change-Id: I3dd4d60e7439541f3ea5f8f44a029d13bc753932
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11461
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>