io-sched: handle exchange with device that belongs to no I/O sched
When an I/O scheduler wants to exchange a device that belongs to no I/O
scheduler yet, the type of the I/O request of the lrs_dev structure is
0 (i.e. `lrs_dev::ld_io_request_type == 0`). In this case, we simply add
the device to the I/O scheduler. No need to claim it from another I/O
scheduler.

This behavior may break the current device repartition between
schedulers but this will be temporary. The next iteration of the
scheduler thread loop will redistribute the devices.

Change-Id: I060db03633b2d525d359ca2d4c9bdc90271e20ea
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11481
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>