io-sched: fix medium search functions
When looking for a medium in a device, the LRS would call the functions
io_sched_search_loaded_medium or io_sched_search_in_use_medium depending
on the context. These functions were modified to only search the I/O
schedulers' list of devices.

This behavior is not correct when using the fair share algorithm. Since
the fair share algorithm may remove devices from those, other threads
may not see all the devices. For example, the communication thread may
not find the device containing a medium that is contained in a release
request.

Furthermore, some I/O schedulers need to know if a medium is loaded on a
device. Since the fair share only gives them a partial view of the
devices, a medium may be loaded on a device that they do not currently
own. Therefore, they also always need to look at each device, not just
the one they own.

This patch reintroduces the function search_in_use_medium and
search_loaded_medium. These functions need to be called with the
complete list of devices.

Since I/O schedulers do not have access to the list of devices, a
pointer to this list is added `io_sched_handle::global_device_list`.

Change-Id: I387848414c281966838a9109d06ff7d41c3ac66e
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11485
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>