io-sched fair share: ignore devices that are not available
The current implementation of the LRS does not remove devices that
are set to failed after an error occured. This means that the global
list of devices `lrs_sched::devices` may contain devices that cannot be
used. Before this patch, they were taken into account when computing the
device repartition.

In the worst case, this can lead to the impossibility for an I/O
scheduler that has only one device to perform any action since it may
receive a failed device.

The function `dev_is_online` has been added to make sure that the fair
share algorithm is able to ignore these devices.

Change-Id: Ie9db66cd953dd83b147d87b3538dba5023680b9c
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11486
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>