lrs: do not unlock the media during read allocation
Prior to this patch, the lrs would unlock the media in
sched_read_alloc_one_medium when no device was found to load this
medium. This behavior assumed that the read I/O scheduler would lock
the medium internally. This assumption is wrong, therefore, this patch
removes the unlock.

We also update acceptance.test to run dir_lock_test last since this test
will insert new directories in the database and remove them at the end.
This leaves phobos in an inconsistent state and makes the test
concurrent_put_get_retry fail.

Change-Id: Iebb5189565ee659d71044daca5ba33184bda8dc3
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11577
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>