io-sched group read: make sure that a medium is not loaded
When searching for a new queue to allocate in grouped_peek_request,
we need to make sure that the medium is not loaded before searching a
compatible drive. If the fair share algorithm is used with group read,
a queue may not have a device associated to it since the device may
belong to another scheduler. In this case, we need to check if the
medium associated with the queue is loaded in a device. If that is the
case, we can try to exchange it with the other I/O scheduler to allocate
this request. If we fail to exchange the device, we simply skip this
queue and try to allocate the next one in the queue table.

Change-Id: I89516cc6beb04a2e9164a826fda0e39dd84f0084
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11579
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>