io-sched group read: fix invalid list traversal
In request_can_be_allocated, a list extra_devices is used to keep track
of free devices that we can try to exchange when we don't have enough
devices for the current allocation. Once we have looked at every
possible device, if we don't have enough devices to allocate the
request, we will loop through extra_devices to try to exchange new
devices with other schedulers.

This patch also:
- removes the function remove_queue_element which did the
  exact same operations as remove_element_from_queue.
- add associate_queue_to_device and remove_queue_from_device to isolate
  these operations and make debugging easier

Change-Id: Ifeb8d92cd30b135ef24c2c236a84f370d612712a
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/11578
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>